requirements.txt
# ClaimDesk AI — Backend Requirements
fastapi
uvicorn[standard]
python-dotenv
twilio
faster-whisper
transformers
torch
chromadb
sentence-transformers
huggingface_hub
rank_bm25
numpy
scipy
pydantic